9479: consolidating 2 endpoints into 1
