+++ shared/src/business/useCaseHelper/stampDisposition/addDraftStampOrderDocketEntryInteractor.js
@@ -1,12 +1,12 @@
 const {
   COURT_ISSUED_EVENT_CODES,
   DOCUMENT_PROCESSING_STATUS_OPTIONS,
-} = require('../entities/EntityConstants');
-const { Case } = require('../entities/cases/Case');
-const { DocketEntry } = require('../entities/DocketEntry');
-const { Message } = require('../entities/Message');
+} = require('../../entities/EntityConstants');
+const { Case } = require('../../entities/cases/Case');
+const { DocketEntry } = require('../../entities/DocketEntry');
+const { Message } = require('../../entities/Message');
 const { orderBy } = require('lodash');
-const { Stamp } = require('../entities/Stamp');
+const { Stamp } = require('../../entities/Stamp');
 
 /**
  * addDraftStampOrderDocketEntryInteractor
