+++ shared/src/business/useCaseHelper/stampDisposition/generateStampedCoversheetInteractor.js
@@ -1,5 +1,7 @@
-const { Case } = require('../entities/cases/Case');
-const { generateCoverSheetData } = require('./generateCoverSheetData');
+const {
+  generateCoverSheetData,
+} = require('../../useCases/generateCoverSheetData');
+const { Case } = require('../../entities/cases/Case');
 
 /**
  * a helper function which creates a coversheet with stampData on it, then returns the new coversheet pdf
