+++ web-api/src/app.js
@@ -69,9 +69,6 @@ const {
   addDeficiencyStatisticLambda,
 } = require('./cases/addDeficiencyStatisticLambda');
 const {
-  addDraftStampOrderDocketEntryLambda,
-} = require('./documents/addDraftStampOrderDocketEntryLambda');
-const {
   addPetitionerToCaseLambda,
 } = require('./cases/addPetitionerToCaseLambda');
 const {
@@ -171,6 +168,9 @@ const {
   generateDocketRecordPdfLambda,
 } = require('./cases/generateDocketRecordPdfLambda');
 const {
+  generateDraftStampOrderLambda,
+} = require('./documents/generateDraftStampOrderLambda');
+const {
   generatePractitionerCaseListPdfLambda,
 } = require('./cases/generatePractitionerCaseListPdfLambda');
 const {
@@ -182,9 +182,6 @@ const {
 const {
   generatePrintablePendingReportLambda,
 } = require('./pendingItems/generatePrintablePendingReportLambda');
-const {
-  generateStampedCoversheetLambda,
-} = require('./documents/generateStampedCoversheetLambda');
 const {
   generateTrialCalendarPdfLambda,
 } = require('./trialSessions/generateTrialCalendarPdfLambda');
@@ -584,8 +581,8 @@ const { validatePdfLambda } = require('./documents/validatePdfLambda');
     lambdaWrapper(addCoversheetLambda),
   );
   app.post(
-    '/case-documents/:docketNumber/:docketEntryId/stamped-coversheet',
-    lambdaWrapper(generateStampedCoversheetLambda),
+    '/case-documents/:docketNumber/:motionDocketEntryId/stamp',
+    lambdaWrapper(generateDraftStampOrderLambda),
   );
   app.post(
     '/case-documents/:docketNumber/:docketEntryId/remove-signature',
@@ -598,10 +595,6 @@ const { validatePdfLambda } = require('./documents/validatePdfLambda');
   app.post(
     '/case-documents/:docketNumber/:docketEntryId/sign',
     lambdaWrapper(saveSignedDocumentLambda),
-  );
-  app.post(
-    '/case-documents/:docketNumber/:docketEntryId/stamp',
-    lambdaWrapper(addDraftStampOrderDocketEntryLambda),
   );
   app.post(
     '/case-documents/:docketNumber/:docketEntryId/serve',
