+++ web-client/src/presenter/actions/completeMotionStampingAction.js
@@ -1,6 +1,8 @@
 import { Stamp } from '../../../../shared/src/business/entities/Stamp';
 import { state } from 'cerebral';
 
+// '/case-documents/:docketNumber/:docketEntryId/stamp',
+
 /**
  * generates an action for completing motion stamping
  *
@@ -15,12 +17,12 @@ export const completeMotionStampingAction = async ({
   get,
   props,
 }) => {
-  const motionDocketEntryID = get(state.pdfForSigning.docketEntryId);
+  const motionDocketEntryId = get(state.pdfForSigning.docketEntryId);
   const { docketNumber } = get(state.caseDetail);
   const parentMessageId = get(state.parentMessageId);
   const stampFormData = get(state.form);
 
-  let newDocketEntryId;
+  const newDocketEntryId = applicationContext.getUniqueId();
 
   const { nameForSigning, nameForSigningLine2 } = get(state.pdfForSigning);
 
@@ -31,24 +33,13 @@ export const completeMotionStampingAction = async ({
     nameForSigningLine2,
   };
 
-  newDocketEntryId = applicationContext.getUniqueId();
-
   await applicationContext
     .getUseCases()
-    .addDraftStampOrderDocketEntryInteractor(applicationContext, {
+    .generateDraftStampOrderInteractor(applicationContext, {
       docketNumber,
       formattedDraftDocumentTitle: props.formattedDraftDocumentTitle,
-      originalDocketEntryId: motionDocketEntryID,
+      motionDocketEntryId,
       parentMessageId,
-      stampData,
-      stampedDocketEntryId: newDocketEntryId,
-    });
-
-  await applicationContext
-    .getUseCases()
-    .generateStampedCoversheetInteractor(applicationContext, {
-      docketEntryId: motionDocketEntryID,
-      docketNumber,
       stampData,
       stampedDocketEntryId: newDocketEntryId,
     });
